Lots of clean up to enable sensible packaging.
Converted the dal into an Arduino build-chain friendly library installable
as a debian package.

Made the iot_tether codebased use this toolchain.  (Build simplification in
fact) Include a prebuilt version of the iot_tether.hex file in a bin directory.
Include python code that uses/demos the iot_tether code.

Shifted the font generation tool into the same location as the DAL.

Tidied up the device directory, in prep for replacement/update.

Worth noting - the Production Data is what was used for producing the
prototype devices. (These are the output files)

Revamped the pycc code base.  (This will be eventually replaced by Pyxie,
but for now this is tidied and packaged up to be instantly usable.  This
revamped version now installs as a debian package and includes two examples
with local makefiles - to demonstrate how to use this package.

Deleted the legacy SWP example "keyword free parser"